% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.lpi.R
\name{check.code}
\alias{check.code}
\title{Identify Unknown Plant Codes and Their Growth Habits and Durations}
\usage{
check.code(code, valid.structure = "^(AF|AG|PF|PG|SH|TR)\\\\d{2,3}$")
}
\arguments{
\item{code}{Character string. The string to check against the unknown plant code patterns}

\item{valid.structure}{Regular expression as a character string. Define what the acceptable structure is for unknown plant codes. Defaults to the AIM format of \code{"^(AF|AG|PF|PG|SH|TR)\\d{2,3}$"}.}
}
\value{
A named list with three values. \code{valid} is logical and indicates if the code was a valid unknown. The remaining values match the style of DIMA's \code{tblSpeciesGrowthHabit}. \code{habit}, a character string, will be \code{NULL} if the code was not valid and either \code{"Woody"} or \code{"Non-woody"} as appropriate if it was valid. \code{subhabit}, a character string, will be \code{NULL} if the code was not valid and \code{"Forb/herb"}, \code{"Graminoid"}, \code{"Shrub"}, or \code{"Tree"} as appropriate if it was valid. \code{duration}, a character string, will be \code{NULL} if the code was not valid and either \code{"Annual"} or \code{"Perennial"} as appropriate if it was valid.
}
\description{
Check a character string to see if it's a valid DIMA/AIM unknown plant code. Additionally, try to assign growth habit and duration for that code if possible.
}
