% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.dima}
\alias{check.dima}
\title{Create a list of data frames documenting errors in plot, LPI, gap, species richness, soil stability, and soil pit forms.}
\usage{
check.dima(dima.tables = NULL, dima.filepath = NULL, dima.filename = NULL,
  all.species = NULL, expected.linecount = 3, write = TRUE)
}
\arguments{
\item{dima.tables}{List of data frames. Optional only if instead providing values for \code{dima.filepath} and \code{dima.filename}. The result of importing one or more DIMA with \code{read.dima()}. Must contain the data frames tblSites, tblPlots, tblLines, tblLPIDetail, tblLPIHeader, tblGapDetail, tblGapHeader, tblSoilStabDetail, tblSoilStabHeader, tblSpecRichDetail, tblSpecRichHeader, tblSoilPits, and tblSoilPitHorizons. Defaults to \code{NULL}.}

\item{dima.filepath}{Character string. Optional only if instead providing a list of data frames as \code{dima.tables}. If provided, must match the full filepath to the folder containing the DIMA or DIMAs to read in. Defaults to \code{NULL}.}

\item{dima.filename}{Character string. Optional only if instead providing a list of data frames as \code{dima.tables}. If provided, must match the full filename or filenames of the DIMA or DIMAs to read in from the folder specified in \code{dima.filepath}. Defaults to \code{NULL}.}

\item{all.species}{Data frame. The full list of species found in the USDA PLANTS database. Must contain the species symbol in the variable "Symbol". Defaults to the result of \code{default.species()}.}

\item{expected.linecount}{Numeric. The expected number of lines recorded on per plot. Defaults to \code{3}.}

\item{write}{Logical. If \code{TRUE} and \code{dima.filepath} is not \code{NULL} then the error data frames will be written as CSV files into that filepath. Defaults to \code{TRUE}.}
}
\description{
Create a list of data frames documenting errors in plot, LPI, gap, species richness, soil stability, and soil pit forms.
}
