% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.gap}
\alias{check.gap}
\title{Create a data frame documenting errors in gap forms}
\usage{
check.gap(detail.table, header.table, sites.plots.lines.meta)
}
\arguments{
\item{detail.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblGapDetail.}

\item{header.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblGapHeader.}

\item{sites.plots.lines.meta}{Data frame. The result of merging the DIMA tables tblSites and tblPlots by the variable SiteKey and that result with the DIMA table tblLines by LineKey.}
}
\description{
Create a data frame documenting errors in gap forms
}
