% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.header}
\alias{check.header}
\title{Check the minimum requirements of a header table}
\usage{
check.header(header.table)
}
\arguments{
\item{header.table}{Data frame. A header table from DIMA, e.g. tblLPIHeader, tblGapHeader, etc.}
}
\description{
Check the minimum requirements of a header table
}
