% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.plot}
\alias{check.plot}
\title{Create a data frame reporting errors in DIMA plot forms}
\usage{
check.plot(site.table, plot.table, line.table, header.table.lpi,
  header.table.gap, header.table.richness, header.table.soilstability,
  header.table.soilpit, expected.linecount = 3)
}
\arguments{
\item{site.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSites. Must contain at least the variables \code{"SiteKey"} and \code{"SiteID"}.}

\item{plot.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblPlots.}

\item{line.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblLines. Must contain at least the variables \code{"PlotKey"}, \code{"LineKey"}, and \code{"LineID"}.}

\item{header.table.lpi}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblLPIHeader. Must contain at least the variables \code{"LineKey"} and \code{"RecKey"}.}

\item{header.table.gap}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblGapHeader. Must contain at least the variables \code{"LineKey"} and \code{"RecKey"}.}

\item{header.table.richness}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSpecRichHeader. Must contain at least the variables \code{"LineKey"} and \code{"RecKey"}.}

\item{header.table.soilstability}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSoilStabHeader. Must contain at least the variables \code{"LineKey"} and \code{"RecKey"}.}

\item{header.table.soilpit}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSoilPits. Must contain at least the variables \code{"PlotKey"} and \code{"SoilKey"}.}

\item{expected.linecount}{Numeric. The expected number of lines recorded on per plot. Defaults to \code{3}.}
}
\description{
Check the plot form tables for errors or incompleteness and return a data frame detailing the errors.
}
