% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.richness}
\alias{check.richness}
\title{Create a data frame documenting errors in species richness forms}
\usage{
check.richness(detail.table, header.table, sites.plots.lines, all.species)
}
\arguments{
\item{detail.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSpecRichDetail.}

\item{header.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSpecRichHeader.}

\item{sites.plots.lines}{Data frame. The result of merging the DIMA tables tblSites and tblPlots by the variable SiteKey and that result with the DIMA table tblLines by LineKey.}

\item{all.species}{Data frame. The full list of species found in the USDA PLANTS database. Must contain the species symbol in the variable "Symbol". Defaults to the result of \code{default.species()}.}
}
\description{
Create a data frame documenting errors in species richness forms
}
