% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check.soilpit}
\alias{check.soilpit}
\title{Create a data frame documenting errors in soil pit forms}
\usage{
check.soilpit(header.table, detail.table, sites.plots)
}
\arguments{
\item{header.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSoilPit.}

\item{detail.table}{Data frame. In DIMA and the output from \code{read.dima()}, this is the table called tblSoilPitHorizons.}

\item{sites.plots.}{Data frame. The result of merging the DIMA tables tblSites and tblPlots by SiteKey.}
}
\description{
Create a data frame documenting errors in soil pit forms
}
