% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.species.R
\name{common.species}
\alias{common.species}
\title{Find the most common species characteristics assigned in DIMAs}
\usage{
common.species(data.path, dima.list)
}
\arguments{
\item{data.path}{A string. The path to the folder containing the DIMA[s].}

\item{dima.list}{Optional character vector. Two or more filenames of DIMAs to read data from in \code{data.path}. If not specified, all DIMAs in the folder will be read from.}
}
\value{
A list with containing data frames \code{frequent} and \code{unique}.
}
\description{
Compare two or more DIMAs against each other to find out what the variability in species characteristic assignment is like and what the most common assignments are. This pulls in data from DIMAs without combining them because species are expected to appear more than once and combined tables will wreak havoc with some \code{merge()} steps. The data frame \code{uniques} contains one entry per combination of species code, growth habit, growth sub habit, and duration and a count of how many databases that combination occurred in. The data frame \code{frequent} contains one entry for each species code that had at least one of those characteristics assigned and the most frequently occurring value for each characteristic. These are considered separately though, so if ARTRW8 shows up in three DIMAs as non-woody perennial forb, a woody perennial shrub, and a woody annual shrub then the entry in the output would call it a woody perennial shrub.
}
