% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{create.errorframe}
\alias{create.errorframe}
\title{Create a data frame documenting errors}
\usage{
create.errorframe(source.df, source.variables = NULL, error.text,
  error.vector, error.component = NULL)
}
\arguments{
\item{source.df}{Data frame. The data frame containing the metadata entries, if relevant, the variable matching \code{error.component} to be used in creating a more specific error text string.}

\item{source.variables}{Option vector of character strings. If \code{source.df} contains extraneous variables, this can be used to list the variables to keep from source.df. If \code{NULL} then all variables will be kept. Defaults to \code{NULL}.}

\item{error.text}{Character string. The error message to add to the variable \code{error} in the output. This will be the same for all entries in the output data frame unless a variable is given in the argument \code{error.component} in which case the values from that variable will be added to the error text separated by " ".}

\item{error.vector}{Logical vector. The logical vector used to slice \code{source.df} down to only entries in which an error occurred. \code{TRUE} values must correspond to an error occurring in \code{source.df}. The vector must be in the same order as \code{source.df} and the same length.}

\item{error.component}{Optional character string matching a variable name in \code{source.df}. If provided, the values from this variable will be added to the end of the \code{error.text} string separated by a " ", allowing for more specific error text, e.g. "Invalid surface code: Rock" instead of just "Invalid surface code". Defaults to \code{NULL}.}
}
\description{
This takes a source data frame with the relevant metadata and potentially a variable to flag as an error, adds a variable with the error message, and restricts the output to only the requested variables and the error message for the entries that had errors.
}
