% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{extract.table}
\alias{extract.table}
\title{Reading in data from DIMA}
\usage{
extract.table(data.path, dima, query)
}
\arguments{
\item{data.path}{A string specifying the folder path containing the DIMA[s].}

\item{dima}{A string specifying the filename with extension of the DIMA to read data from in \code{data.path}.}

\item{query}{An named character vector of one or more SQL queries. Value names should follow the same pattern as "gap", "lpi", "species.inventory".}
}
\description{
This will read in data from a DIMA according to the SQL query or queries provided. The output is a named list of query results.
}
