% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R, R/height_new.R
\name{mean.height}
\alias{mean.height}
\alias{mean.height}
\title{Calculate the vegetation height}
\usage{
\method{mean}{height}(lpi.tall, by.line = FALSE, tall = FALSE,
  by.year = FALSE, by.max = FALSE, omit.zero = FALSE, ...)

\method{mean}{height}(lpi.tall, by.line = FALSE, tall = FALSE,
  by.year = FALSE, by.max = FALSE, omit.zero = FALSE, ...)
}
\arguments{
\item{lpi.tall}{A tall/long-format data frame. Use the data frame \code{"height"} from the \code{gather.lpi()} output.}

\item{by.line}{Logical. If \code{TRUE} then results will be reported further grouped by line using the \code{LineID} and \code{LineKey} fields from the data forms. Defaults to \code{FALSE}.}

\item{tall}{Logical. If \code{TRUE} then the returned data frame will be tall rather than wide and will not have observations for non-existent values e.g., if no data fell into a group on a plot, there will be no row for that group on that plot. Defaults to \code{FALSE}.}

\item{by.year}{Logical. If \code{TRUE} then results will be reported further grouped by year using the \code{DateModified} field from the data forms. Defaults to \code{FALSE}.}

\item{omit.zero}{Logical. If \code{TRUE} the results omit height measurements of \code {0}}

\item{...}{One or more bare variable name from \code{lpi.tall} to calculate percent cover for, e.g. \code{GrowthHabitSub} to calculate percent cover by growth habits or \code{GrowthHabitSub, Duration} to calculate percent cover for categories like perennial forbs, annual graminoids, etc.}

\item{hit}{Character string. If \code{"any"} then percent cover will be calculated using any hit in the canopy column (so a single pin drop record may be counted more than once if it had hits that corresponded to different groups). If \code{"first"} then only the first canopy hit at a pin drop will be used to calculate cover. Defaults to \code{"any"}.}

\item{by.species}{Logical. If \code{TRUE} the results will be reported by species and any growth habit assignments in \code {...}}

\item{lpi.tall}{A tall/long-format data frame. Use the data frame \code{"height"} from the \code{gather.lpi()} output.}

\item{...}{One or more bare variable name from \code{lpi.tall} to calculate percent cover for, e.g. \code{GrowthHabitSub} to calculate percent cover by growth habits or \code{GrowthHabitSub, Duration} to calculate percent cover for categories like perennial forbs, annual graminoids, etc.}

\item{tall}{Logical. If \code{TRUE} then the returned data frame will be tall rather than wide and will not have observations for non-existent values e.g., if no data fell into a group on a plot, there will be no row for that group on that plot. Defaults to \code{FALSE}.}

\item{by.year}{Logical. If \code{TRUE} then results will be reported further grouped by year using the \code{DateModified} field from the data forms. Defaults to \code{FALSE}.}

\item{by.line}{Logical. If \code{TRUE} then results will be reported further grouped by line using the \code{LineID} and \code{LineKey} fields from the data forms. Defaults to \code{FALSE}.}
}
\description{
Calculate the vegetation height
}
