% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamerge.R
\name{metamerge}
\alias{metamerge}
\title{Add metadata to form data}
\usage{
metamerge(dima.tables = list(), form = "")
}
\arguments{
\item{dima.tables}{A list of data frames. The output from \code{read.dima()} can be provided directly as this argument.}

\item{form}{A string specifying which form to work with. This can either be the form name in the background tables of DIMA or the human version, e.g. \code{"CanopyGap"}, \code{"Canopy Gap"}, \code{"LPI"}, \code{"Line Point Intercept"}, \code{"Species Richness"}, etc. This is case insensitive.}
}
\value{
A data frame created by merging \code{tblSites}, \code{tblPlots}, \code{tblLines}, and the header and detail tables for the selected form.
}
\description{
For a given form, combine the form's header and detail data then merge with site, plot, and line data
}
