% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.gdbtable}
\alias{read.gdbtable}
\title{Read Tables From ESRI Geodatabases}
\usage{
read.gdbtable(dsn, tablename, overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}

\item{tablename}{Character string. The exact name of the table in the file geodatabase being read in.}

\item{overwrite}{Logical. If \code{TRUE} then the intermediate step of writing the table to a temporary file as an ESRI shapefile can overwrite an existing shapefile if they share a filepath and filename. Defaults to \code{TRUE}.}

\item{verbose}{Logical. Passed to \code{gdalUtils::ogr2ogr(verbose)}. Verbose mode can be very helpful for troubleshooting. Defaults to \code{FALSE}.}
}
\value{
Data frame
}
\description{
A workaround using \code{gdalUtils::ogr2ogr()} and \code{foreign::read.dbf()} to read tables without geometry from ESRI geodatabases into R environments as data frames.
}
