% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.dima}
\alias{read.dima}
\title{Reading in data from multiple DIMAs}
\usage{
read.dima(data.path, dima.list = NULL, all.tables = FALSE, gap = FALSE,
  lpi = FALSE, species = FALSE, species.inventory = FALSE,
  stability = FALSE, stability.tidy = FALSE, custom.query = NULL,
  combine = TRUE)
}
\arguments{
\item{data.path}{Character string. The path to the folder containing the DIMA[s].}

\item{dima.list}{Optional character vector. One or more filenames of DIMAs to read data from in \code{data.path}. If not specified, all DIMAs in the folder will be read from.}

\item{gap}{Logical. If \code{TRUE} then canopy gap data will be read in with a default SQL query. Defaults to \code{FALSE}.}

\item{lpi}{Logical. If \code{TRUE} then line-point intercept data will be read in with a default SQL query. Defaults to \code{FALSE}.}

\item{species}{Logical. If \code{TRUE} then species data will be read in with a default SQL query. Defaults to \code{FALSE}.}

\item{species.inventory}{Logical. If \code{TRUE} then species inventory data will be read in with a default SQL query. Defaults to \code{FALSE}.}

\item{stability}{Logical. If \code{TRUE} then soil stability data will be read in with a default SQL query. Defaults to \code{FALSE}.}

\item{stability.tidy}{Logical. \code{TRUE} then soil stability data will be reformatted from the format in DIMA into a tidy data frame. Defaults to \code{FALSE}.}

\item{custom.query}{An optional named character vector of one or more SQL queries. Value names should follow the same pattern as "gap", "lpi", "species.inventory".}

\item{combine}{Logical. If \code{TRUE} then the output will be a named list of data frames, one for each SQL query. The data frames will contain all of the relevant data from all of the DIMAs read from. If \code{FALSE} then the output will be a named list of per-DIMA named lists of query result data frames. Defaults to \code{TRUE}.}

\item{all}{Logical. If \code{TRUE} then the contents of all tables in the database will be read in. Defaults to \code{FALSE}.}
}
\description{
This will read in data from one or more DIMAs according to the SQL queries requested and includes default SQL queries for canopy gap, line-point intercept, soil stability, and species inventory data. The output is either a list of lists of query results named with the source filename[s] or a list of query results combined from all sources named with the query name[s].
}
